+++ docs/plugins/site_builder.rb
@@ -1,0 +1,3 @@
+class SiteBuilder < Bridgetown::Builder
+  # write builders which subclass SiteBuilder in plugins/builder
+end