+++ docs/postcss.config.js
@@ -1,0 +1,7 @@
+const plugins = {
+  "autoprefixer": {},
+}
+
+module.exports = {
+  plugins
+};