+++ src/clickHandler.ts
@@ -1,4 +1,4 @@
-import { Utils } from './utils'
+import { Misc } from './utils/misc'
 import { match } from './utils/dom'
 import { EventQueryInterface } from './types'
 
@@ -16,11 +16,15 @@ export class ClickHandler {
     ]
   }
 
+  get name (): string {
+    return ClickHandler.name
+  }
+
   connect (): void {
     ClickHandler.queries.forEach((obj) => {
       obj.selectors.forEach((selector) => {
         document.querySelectorAll(selector).forEach((element) => {
-          element.addEventListener(obj.event, Utils.preventInsignificantClick as EventListener)
+          element.addEventListener(obj.event, Misc.preventInsignificantClick as EventListener)
         })
       })
     })
@@ -30,22 +34,22 @@ export class ClickHandler {
     ClickHandler.queries.forEach((obj) => {
       obj.selectors.forEach((selector) => {
         document.querySelectorAll(selector).forEach((element) => {
-          element.removeEventListener(obj.event, Utils.preventInsignificantClick as EventListener)
+          element.removeEventListener(obj.event, Misc.preventInsignificantClick as EventListener)
         })
       })
     })
   }
 
-  observerCallback (nodeList: NodeList): void {
+  observerCallback (nodeList: Node[]): void {
     ClickHandler.queries.forEach((obj) => {
       obj.selectors.forEach((selector) => {
         nodeList.forEach((node) => {
           if (match(node, { selector })) {
-            node.addEventListener(obj.event, Utils.preventInsignificantClick as EventListener)
+            node.addEventListener(obj.event, Misc.preventInsignificantClick as EventListener)
           }
 
           if (node instanceof Element) {
-            node.querySelectorAll(selector).forEach((el) => el.addEventListener(obj.event, Utils.preventInsignificantClick as EventListener))
+            node.querySelectorAll(selector).forEach((el) => el.addEventListener(obj.event, Misc.preventInsignificantClick as EventListener))
           }
         })
       })