+++ src/types.ts
@@ -1,11 +1,11 @@
 import { FetchResponse } from './http/fetchResponse'
 import { FetchRequest } from './http/fetchRequest'
-
+export type Locateable = URL | string
 export type AddOrRemoveListeners = 'addEventListener' | 'removeEventListener'
 export type Submitter = HTMLInputElement | HTMLButtonElement
 
 export interface EventQueryInterface {
-  event: 'click' | 'change' | 'submit'
+  event: string
   selectors: string[]
 }
 
@@ -57,12 +57,23 @@ export interface MimeTypeInterface {
 
 export interface MrujsPluginInterface {
   name: string
+  initialize?: () => void
   connect: () => void
   disconnect: () => void
+  observerCallback?: (addedNodes: Node[]) => void
 }
 
 export interface ExposedUtilsInterface {
   match: Function
   FetchRequest: FetchRequest['constructor']
   FetchResponse: FetchResponse['constructor']
+}
+
+export interface SnapshotCacheInterface {
+  size: number
+  keys: string[]
+  snapshots: Record<string, unknown>
+  put: (location: URL, snapshot: string) => void
+  has: (location: URL) => boolean
+  [key: string]: unknown
 }