+++ test/rails/dummy/app/javascript/packs/application.js
@@ -1,13 +1,13 @@
-// This file is automatically compiled by Webpack, along with any other files
-// present in this directory. You're encouraged to place your actual application logic in
-// a relevant structure within app/javascript and only use these pack files to reference
-// that code so it'll be compiled.
-
 import Rails from 'mrujs'
-// import Rails from '@rails/ujs'
-import Turbolinks from 'turbolinks'
-import * as ActiveStorage from '@rails/activestorage'
+
+(async () => {
+  if (window.useTurbo === true) {
+    const Turbo = await import('@hotwired/turbo')
+    window.Turbo = Turbo
+  } else {
+    const turbolinks = await import('turbolinks')
+    turbolinks.default.start()
+  }
+})()
 
 Rails.start()
-Turbolinks.start()
-ActiveStorage.start()