+++ test/rails/dummy/test/system/turbo_navigation_adapter_test.rb
@@ -1,0 +1,11 @@
+require "turbo_test_case"
+
+require_relative "./navigation_adapter_tests"
+
+class TurboNavigationAdapterTest < TurboTestCase
+  include NavigationAdapterTests
+
+  setup do
+    turbo_visit root_path
+  end
+end