+++ test/rails/dummy/test/system/turbo_remote_forms_test.rb
@@ -1,0 +1,15 @@
+require "application_system_test_case"
+
+require_relative "./remote_form_tests"
+
+class TurboRemoteFormsTest < TurboTestCase
+  include ::RemoteFormTests
+
+  setup do
+    before_visit
+    @post = posts(:one)
+    visit posts_path
+    after_visit
+  end
+end
+