+++ test/rails/dummy/test/system/turbo_remote_links_test.rb
@@ -1,0 +1,11 @@
+require "turbo_test_case"
+
+require_relative "./remote_link_tests"
+
+class TurboRemoteLinksTest < TurboTestCase
+  include ::RemoteLinkTests
+
+  def setup
+    turbo_visit remote_links_path
+  end
+end