+++ test/rails/dummy/test/system/turbolinks_remote_forms_test.rb
@@ -1,0 +1,13 @@
+require "turbolinks_test_case"
+
+require_relative "./remote_form_tests"
+
+class TurbolinksRemoteFormsTest < TurbolinksTestCase
+  include ::RemoteFormTests
+
+  setup do
+    @post = posts(:one)
+    turbolinks_visit posts_path
+  end
+end
+